edit client details modal as separate component for reuse
